Use calculated native version
Instead of using a manual version for the native Library,
the version is generated from the hashes of the native sources.
This way, the version is different when the version is different
when the native libraries change and the caching of the directory
to where the unpacking happens is safe to use even for snapshot
versions.