Use a single FSEventStream
Previously we started an event stream for each individual directory hierarchy we watched. This seems to be overly expensive and there's also a limit of ~1000 event streams per machine that we might hit. So instead we are now using a single event stream that watches multiple directory hierarchies.

This makes adding/removing watches slightly slower.

There's also a change in behavior wrt watching the same directory both directly and via a symlink: in this case changes are not reported twice with the same path anymore, which is arguably better.