Use the right host_flavor_t value for returning host_info64_t data
With HOST_VM_INFO, the syscall only fills in parts of the structure.
With HOST_VM_INFO64, all of the structure is filled in.

This caused very strange "stack corrupt" looking problems where the
exact amount of data filled into the structure seemed to depend on
its alignment on the stack. Adding dummy values to the stack would
cause more of the vm_statistics64_data_t structure to be filled in.