Swap Gradle's name for the 64-bit ARM architecture with native-platform's name
We have a mismatch between Gradle's native plugins and native-platform.

In Gradle's native plugin, we consider arm-v8 and arm64 to be the 64-bit
ARM architecture names.

In native-platform, we recognize aarch64 and arm64 as the 64-bit ARM
architecture names.

We'll need to fix this mismatch upstream and update native-platform later.

I think aarch64 is the better name.